+++ src/main/java/com/authlete/common/api/AuthleteApiImpl.java
@@ -56,6 +56,12 @@ import com.authlete.common.dto.ClientRegistrationResponse;
 import com.authlete.common.dto.ClientSecretRefreshResponse;
 import com.authlete.common.dto.ClientSecretUpdateRequest;
 import com.authlete.common.dto.ClientSecretUpdateResponse;
+import com.authlete.common.dto.DeviceAuthorizationRequest;
+import com.authlete.common.dto.DeviceAuthorizationResponse;
+import com.authlete.common.dto.DeviceCompleteRequest;
+import com.authlete.common.dto.DeviceCompleteResponse;
+import com.authlete.common.dto.DeviceVerificationRequest;
+import com.authlete.common.dto.DeviceVerificationResponse;
 import com.authlete.common.dto.GrantedScopesGetResponse;
 import com.authlete.common.dto.IntrospectionRequest;
 import com.authlete.common.dto.IntrospectionResponse;
@@ -152,6 +158,9 @@ class AuthleteApiImpl implements AuthleteApi
     private static final String BC_AUTHENTICATION_COMPLETE_API_PATH    = "/api/backchannel/authentication/complete";
     private static final String BC_AUTHENTICATION_FAIL_API_PATH        = "/api/backchannel/authentication/fail";
     private static final String BC_AUTHENTICATION_ISSUE_API_PATH       = "/api/backchannel/authentication/issue";
+    private static final String DEVICE_AUTHORIZATION_API_PATH          = "/api/device/authorization";
+    private static final String DEVICE_COMPLETE_API_PATH               = "/api/device/complete";
+    private static final String DEVICE_VERIFICATION_API_PATH           = "/api/device/verification";
 
 
     private final String mBaseUrl;
@@ -1442,5 +1451,35 @@ class AuthleteApiImpl implements AuthleteApi
         return callServicePostApi(
                 BC_AUTHENTICATION_COMPLETE_API_PATH, request,
                 BackchannelAuthenticationCompleteResponse.class);
+    }
+
+
+    @Override
+    public DeviceAuthorizationResponse deviceAuthorization(
+            DeviceAuthorizationRequest request) throws AuthleteApiException
+    {
+        return callServicePostApi(
+                DEVICE_AUTHORIZATION_API_PATH, request,
+                DeviceAuthorizationResponse.class);
+    }
+
+
+    @Override
+    public DeviceCompleteResponse deviceComplete(
+            DeviceCompleteRequest request) throws AuthleteApiException
+    {
+        return callServicePostApi(
+                DEVICE_COMPLETE_API_PATH, request,
+                DeviceCompleteResponse.class);
+    }
+
+
+    @Override
+    public DeviceVerificationResponse deviceVerification(
+            DeviceVerificationRequest request) throws AuthleteApiException
+    {
+        return callServicePostApi(
+                DEVICE_VERIFICATION_API_PATH, request,
+                DeviceVerificationResponse.class);
     }
 }